refactor: resolve some issues
