+++ truelearn/preprocessing/__init__.py
@@ -1,14 +1,12 @@
 """Handles the preprocessing tasks required to use the classifiers."""
 
-from .utils import get_values_mean, get_values_sample_std, get_values_population_std
-from .wikifier import page_rank_as_key, cosine_as_key, Annotation, Wikifier
+from ._utils import get_values_mean, get_values_sample_std, get_values_population_std
+from ._wikifier import Annotation, Wikifier
 
 __all__ = [
     "get_values_mean",
     "get_values_sample_std",
     "get_values_population_std",
-    "page_rank_as_key",
-    "cosine_as_key",
     "Annotation",
     "Wikifier",
 ]
