+++ truelearn/preprocessing/_wikifier.py
@@ -1,35 +1,11 @@
 import itertools
-from typing import Optional, Union, Dict, List, Iterable, Callable, cast
+from typing import Optional, Union, Dict, List, Iterable, cast
 from urllib import parse, request
 
 import orjson
 
 Annotation = Dict[str, Union[str, float, None]]
 WikifierResponse = Dict[str, Union[List[Annotation], List[str]]]
-
-
-def page_rank_as_key(annotation: Annotation) -> float:
-    """Return page rank from annotation.
-
-    Args:
-        annotation: An annotation from the wikifier API.
-
-    Returns:
-        A float indicating the page rank of the item.
-    """
-    return cast(float, annotation["pageRank"])
-
-
-def cosine_as_key(annotation: Annotation) -> float:
-    """Return cosine from annotation.
-
-    Args:
-        annotation: An annotation from the wikifier API.
-
-    Returns:
-        A float indicating the cosine of the item.
-    """
-    return cast(float, annotation["cosine"])
 
 
 class Wikifier:
@@ -52,7 +28,7 @@ class Wikifier:
         df_ignore: int = 50,
         words_ignore: int = 50,
         top_n: Optional[int] = None,
-        key_fn: Callable[[Annotation], Union[float, str]] = cosine_as_key,
+        key_fn: str = "cosine",
     ) -> List[Annotation]:
         """Annotate input text using the Wikifier API.
 
@@ -75,9 +51,9 @@ class Wikifier:
                 via key_fn. If None, return all the annotations.
                 Defaults to None.
             key_fn:
-                A function that takes in an annotation and returns a value that
-                can be used to sort different annotations.
-                Defaults to cosine_as_key, which sorts the annotations based on
+                A string representing the key function that is used when sorting
+                the annotations. The allowed values are "cosine" and "pagerank".
+                Defaults to cosine, which sorts the annotations based on
                 their cosine values.
 
         Returns:
@@ -87,14 +63,15 @@ class Wikifier:
 
         Raises:
             ValueError:
-                The response from Wikifier contained an error message
-                or the API key is not valid.
+                1) The response from Wikifier contained an error message.
+                2) The API key is not valid.
+                3) The key_fn is neither cosine nor pagerank.
             urllib.error.HTTPError:
                 The HTTP request returns a status code representing
                 an error.
         """
         resp = self.__make_wikifier_request(text, df_ignore, words_ignore)
-        return self.__format_wikifier_response(resp, top_n, key_fn)
+        return Wikifier.__format_wikifier_response(resp, top_n, key_fn)
 
     def __make_wikifier_request(
         self, text: str, df_ignore: int, words_ignore: int
@@ -143,11 +120,11 @@ class Wikifier:
                 raise ValueError(resp["status"])
             return resp
 
+    @staticmethod
     def __format_wikifier_response(
-        self,
         resp: WikifierResponse,
         top_n: Optional[int],
-        key_fn: Callable[[Annotation], Union[float, str]],
+        key_fn: str,
     ) -> List[Annotation]:
         """Extract annotations from response object.
 
@@ -162,13 +139,20 @@ class Wikifier:
                 only return the top 5 annotations sorted by keys extracted
                 via key_fn.
             key_fn:
-                A function that takes in an annotation and returns a value that
-                can be used to sort different annotations.
+                A string representing the key function that is used when sorting
+                the annotations. The allowed values are "cosine" and "pagerank".
+                Defaults to cosine, which sorts the annotations based on
+                their cosine values.
 
         Returns:
             The list of annotations obtained from the Wikifier API, sorted by
             using the key extracted from key_fn function.
         """
+        if key_fn not in ("cosine", "pagerank"):
+            raise ValueError(
+                "key_fn is expected to be cosine or pagerank."
+                f" Got key_fn={key_fn} instead."
+            )
 
         def __restructure_annotation(
             annotation: Annotation,
@@ -181,12 +165,36 @@ class Wikifier:
                 "wikiDataItemId": annotation["wikiDataItemId"],
             }
 
+        def page_rank_as_key(annotation: Annotation) -> float:
+            """Return page rank from annotation.
+
+            This method could be used to change how the annotations
+
+            Args:
+                annotation: An annotation from the wikifier API.
+
+            Returns:
+                A float indicating the page rank of the item.
+            """
+            return cast(float, annotation["pageRank"])
+
+        def cosine_as_key(annotation: Annotation) -> float:
+            """Return cosine from annotation.
+
+            Args:
+                annotation: An annotation from the wikifier API.
+
+            Returns:
+                A float indicating the cosine of the item.
+            """
+            return cast(float, annotation["cosine"])
+
         annotations = sorted(
             map(
                 __restructure_annotation,
                 cast(Iterable[Annotation], resp["annotations"]),
             ),
-            key=key_fn,
+            key=cosine_as_key if key_fn == "cosine" else page_rank_as_key,
             reverse=True,
         )
